refactor: move home related components to home module
