+++ src/app/(app)/(home)/layout.tsx
@@ -1,10 +1,9 @@
 import { Suspense } from "react";
 
+import { Footer } from "@/modules/home/ui/components/footer";
+import { Navbar } from "@/modules/home/ui/components/navbar";
+import { SearchFilters, SearchFiltersLoadingSkeleton } from "@/modules/home/ui/components/search-filters";
 import { HydrateClient, prefetch, trpc } from "@/trpc/server";
-
-import { Footer } from "./footer";
-import { Navbar } from "./navbar";
-import { SearchFilters, SearchFiltersLoadingSkeleton } from "./search-filters";
 
 interface Props {
   children: React.ReactNode;
